feat: only place one up and down stair in the hells
This changes the level generator to only place one stair in each hell
branch, as a complement to the previous change to level generation size.

This commit will currenlty break crawl due to the number of vaults that
place disconnected stair bubbles in Hell. Follow-up commits will be
adjusting hell vaults to account for all level generator changes.