feat: new monster, putrid mouth
Putrid mouths are a new Tar monster, meant to be a more substatial (ha)
wraith genus monster to fill out the branch flavour set of threats. They
are based on Chinese mythology around hungry ghosts and re-use the old
hungry ghost tile since it was so cool. As a monster they aren't the
toughest of Tar threat, but the mix of miasma, poison, and warning cry
should combine with the small levels to prevent a player who mismanaged
the encounter from recovering before reenforcements arrive.