squash! feat: persistent hell effects: rot tartarus (Lici)
Replace the half-regen-and-mp-regen effect with half willpower. Halved
regen rates increase the resting time, and do have some tactical impact
in longer fights. But if we want resting to be more punishing, the timed
effects themselves can be scaled. The purpose of the branch-wide effects
is to provide a tactical change in every fight.

Ereshkigal has paralysis, and the new Tar monster set has some will
checking monsters. Lets make it harder for the player to simply Will++++
item their way out of these concerns.