feat: new monster, nargun
Without cancellation pots, status effects are much more deadly in
Cocytus. Nargun are modelled after the Australian mythic stone
creatures, and are gargoyle-esque: they hit hard, have big AC, and a
threatening spell but weak HP. In this case they bring Petrify to the
mix, along with AF_VULN and a big HD to help it land. They are lrdable
and treated as icy.